Move the comparator to an upper level.
