+++ server/src/main/java/io/spine/server/delivery/CatchUpStation.java
@@ -22,12 +22,8 @@ package io.spine.server.delivery;
 
 import com.google.protobuf.Duration;
 import com.google.protobuf.util.Durations;
-import io.spine.core.Event;
 import io.spine.server.delivery.event.CatchUpStarted;
-import io.spine.server.event.EventComparator;
-import io.spine.type.TypeUrl;
 
-import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.Comparator;
@@ -271,39 +267,6 @@ final class CatchUpStation extends Station {
                     default:
                         // Skip the message.
                 }
-            }
-        }
-    }
-
-    /**
-     * The comparator which sorts the messages chronologically, but ensures that if there is
-     * a {@link CatchUpStarted} event in the sorted batch, it goes on top.
-     */
-    private static final class CatchUpMessageComparator
-            implements Comparator<InboxMessage>, Serializable {
-
-        private static final long serialVersionUID = 0L;
-        private static final TypeUrl CATCH_UP_STARTED =
-                TypeUrl.from(CatchUpStarted.getDescriptor());
-
-        @Override
-        public int compare(InboxMessage m1, InboxMessage m2) {
-            if (m1.hasEvent() && m2.hasEvent()) {
-                Event e1 = m1.getEvent();
-                String typeOfFirst = e1.getMessage()
-                                       .getTypeUrl();
-                if (typeOfFirst.equals(CATCH_UP_STARTED.toString())) {
-                    return -1;
-                }
-                Event e2 = m2.getEvent();
-                String typeOfSecond = e2.getMessage()
-                                        .getTypeUrl();
-                if (typeOfSecond.equals(CATCH_UP_STARTED.toString())) {
-                    return 1;
-                }
-                return EventComparator.chronologically.compare(e1, e2);
-            } else {
-                return InboxMessageComparator.chronologically.compare(m1, m2);
             }
         }
     }
